requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coinpaprika==0.1.0
fastapi==0.63.0
h11==0.12.0
idna==2.10
numpy==1.19.5
pandas==1.2.0
pydantic==1.7.3
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.13.3
